Refactor for ASTM updates
With ASTM ballot coming soon and changes for IETF inlined into
their standard this commit updates a lot of the text to reflect
the changes needed to line up with ASTM.

The entire document has been nearly rewritten due to this change
with a lot of moving pieces.

It is expected that most Appendix items will become drafts of
their own (FEC) or be placed in more appropriate documents
(Attestations).